refactor: Factor the user struct out into its own module
